requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
scikit-image>=0.17.2
matplotlib>=3.3.0
pillow>=8.0.0 